# docs/swagger.yaml
basePath: /
definitions:
...
  category.Category:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      parent_id:
        type: integer
    type: object
  category.CategoryTree:
    properties:
      children:
        items:
          $ref: '#/definitions/category.CategoryTree'
        type: array
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
    type: object
  category.CreateRequest:
    properties:
      is_active:
        type: boolean
      name:
        maxLength: 255
        minLength: 1
        type: string
      parent_id:
        type: integer
    required:
    - name
    type: object
  category.ListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/category.Category'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  category.TreeResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/category.CategoryTree'
        type: array
    type: object
  category.UpdateRequest:
    properties:
      is_active:
        type: boolean
      name:
        maxLength: 255
        minLength: 1
        type: string
      parent_id:
        type: integer
    type: object
...
  response.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  response.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.APIError'
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.Meta:
    properties:
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      timestamp:
        description: Useful for frontend debugging / UX
        type: string
    type: object
  response.PaginationMeta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
info:
  contact: {}
  description: Market backend (products, stock, income, sales)
  title: Hezzet Market API
  version: "1.0"
paths:
...
  /categories:
    get:
      description: Get paginated list of active categories with optional search
      parameters:
      - description: Page (default 1)
        in: query
        name: page
        type: integer
      - description: Limit (default 10, max 100)
        in: query
        name: limit
        type: integer
      - description: Skip (legacy, default 0). If provided, overrides page/offset.
        in: query
        name: skip
        type: integer
      - description: Search by name
        in: query
        name: search
        type: string
      - description: Order by field (name, created_at)
        enum:
        - name
        - created_at
        in: query
        name: order_by
        type: string
      - description: Order direction (asc/desc)
        enum:
        - asc
        - desc
        in: query
        name: order_direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/category.ListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new category with optional parent
      parameters:
      - description: Category data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/category.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/category.Category'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - Categories
  /categories/{id}:
    delete:
      description: Soft delete category by setting is_active=false
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete category (soft)
      tags:
      - Categories
    get:
      description: Get a single category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/category.Category'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get category by ID
      tags:
      - Categories
    patch:
      consumes:
      - application/json
      description: Update category fields (name, parent_id, is_active)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/category.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/category.Category'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - Categories
  /categories/tree:
    get:
      description: Get hierarchical tree of all active categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/category.TreeResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get category tree
      tags:
      - Categories
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"